+++ app/utils/defaultMicrobreakIdeas.js
@@ -20,5 +20,6 @@ module.exports = [
   { data: 'Raise your right arm, stretch it over your head to the left, and hold for 10 seconds. Repeat on the other side.', enabled: true },
   { data: 'With your right hand, grab each finger of your left hand in turn and squeeze. Repeat on the other side.', enabled: true },
   { data: 'Stand up and do a lunge. Hold for 10 seconds, then do the other leg.', enabled: true },
-  { data: 'Close your eyes and simply notice whatever arises in current moment, without judgement.', enabled: true }
+  { data: 'Close your eyes and simply notice whatever arises in current moment, without judgement.', enabled: true },
+  { data: 'One should focus every 20 minutes for 20 seconds on an object that is 20 feet distance.', enabled: true }
 ]