+++ app/break.js
@@ -24,7 +24,7 @@ ipcRenderer.on('breakIdea', (event, message) => {
   breakText.innerHTML = message[1]
 })
 
-ipcRenderer.on('progress', (event, started, duration, strictMode, postpone, postponePercent) => {
+ipcRenderer.on('progress', (event, started, duration, strictMode, postpone, postponePercent, hardcoreMode) => {
   const progress = document.querySelector('#progress')
   const progressTime = document.querySelector('#progress-time')
   const postponeElement = document.querySelector('#postpone')
@@ -33,11 +33,10 @@ ipcRenderer.on('progress', (event, started, duration, strictMode, postpone, post
   window.setInterval(() => {
     if (Date.now() - started < duration) {
       const passedPercent = (Date.now() - started) / duration * 100
-      Utils.canSkip(strictMode, postpone, passedPercent, postponePercent)
       postponeElement.style.display =
-        Utils.canPostpone(postpone, passedPercent, postponePercent) ? 'flex' : 'none'
+        Utils.canPostpone(postpone, passedPercent, postponePercent, hardcoreMode) ? 'flex' : 'none'
       closeElement.style.display =
-        Utils.canSkip(strictMode, postpone, passedPercent, postponePercent) ? 'flex' : 'none'
+        Utils.canSkip(strictMode, postpone, passedPercent, postponePercent, hardcoreMode) ? 'flex' : 'none'
       progress.value = (100 - passedPercent) * progress.max / 100
       progressTime.innerHTML = Utils.formatTimeRemaining(Math.trunc(duration - Date.now() + started))
     }