+++ app/main.js
@@ -337,7 +337,7 @@ function startMicrobreak () {
     breakPlanner.postponesNumber < postponesLimit && postponesLimit > 0
 
   if (!strictMode || postponable) {
-    globalShortcut.register('CommandOrControl+X', () => {
+    globalShortcut.register(settings.get('endBreakShortcut'), () => {
       const passedPercent = (Date.now() - startTime) / breakDuration * 100
       if (Utils.canPostpone(postponable, passedPercent, postponableDurationPercent)) {
         postponeMicrobreak()
@@ -399,7 +399,7 @@ function startMicrobreak () {
       }
       microbreakWinLocal.webContents.send('microbreakIdea', idea)
       microbreakWinLocal.webContents.send('progress', startTime,
-        breakDuration, strictMode, postponable, postponableDurationPercent)
+        breakDuration, strictMode, postponable, postponableDurationPercent, settings.get('endBreakShortcut'))
       microbreakWinLocal.setAlwaysOnTop(true)
     })
     microbreakWinLocal.loadURL(modalPath)
@@ -441,7 +441,7 @@ function startBreak () {
     breakPlanner.postponesNumber < postponesLimit && postponesLimit > 0
 
   if (!strictMode || postponable) {
-    globalShortcut.register('CommandOrControl+X', () => {
+    globalShortcut.register(settings.get('endBreakShortcut'), () => {
       const passedPercent = (Date.now() - startTime) / breakDuration * 100
       if (Utils.canPostpone(postponable, passedPercent, postponableDurationPercent)) {
         postponeBreak()
@@ -503,7 +503,7 @@ function startBreak () {
       }
       breakWinLocal.webContents.send('breakIdea', idea)
       breakWinLocal.webContents.send('progress', startTime,
-        breakDuration, strictMode, postponable, postponableDurationPercent)
+        breakDuration, strictMode, postponable, postponableDurationPercent, settings.get('endBreakShortcut'))
       breakWinLocal.setAlwaysOnTop(true)
     })
     breakWinLocal.loadURL(modalPath)
@@ -524,7 +524,7 @@ function startBreak () {
 }
 
 function breakComplete (shouldPlaySound, windows) {
-  globalShortcut.unregister('CommandOrControl+X')
+  globalShortcut.unregister(settings.get('endBreakShortcut'))
   if (shouldPlaySound && !settings.get('silentNotifications')) {
     processWin.webContents.send('playSound', settings.get('audio'), settings.get('volume'))
   }