+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Added Hungarian translation
 
+### Changed
+- update Chinese (Traditional) translation (#590)
+
 ## [0.99.4] - 2020-05-15
 ### Fixed
 - checkbox across macOS, Windows, Linux